created extension method GetSyntaxIdentifierForBaseType
